Config VDF: Make getVdfSection friendlier
Can now just take the block name, either with or without quotes,
and search for it. Assumes the end pattern will be a closing brace.

We can also guess the indentation level of the given block, based
on the first occurance of the block name. Optionally, the indent
level can be passed if a specific block is desired.